add a couple of 'wait for jobs' methods
These let you wait for a list of jobs (job IDs or actual job
dicts) to be complete, and for the jobs for a specified BUILD
to appear and to be complete. These are things that various
Fedora scheduling/reporting tools need to do, and putting them
in the client library seems reasonable.