wait_build_jobs: don't use wait_jobs, allow de-duping
wait_build_jobs can actually be a lot more efficient than
wait_jobs, because it gets all the info it needs with a
single API call. wait_jobs can't do this (yet) because you
can't do a single API query on a list of job IDs (yet).

This also has the benefit that it will keep up when a job is
restarted for some reason (as it re-runs the 'find all jobs
for BUILD' query each time it wakes up).

Also, move de-duping here (from the Fedora 'check-compose'
script), because other consumers may well want to use it.
Ideally I'd actually want to put this in openQA itself, as
it's more or less duplicating some functionality that already
exists there but cannot be used by anything else (it's baked
into a particular view in the web UI).

This basically tries to filter out duplicate test runs in the
list of job dicts it's passed. Following openQA's overview
function, it uses a key of DISTRI, VERSION, BUILD, TEST, FLAVOR,
ARCH and MACHINE; two jobs with all those values identical will
be considered 'duplicates' and only the newest will be
returned.