rejig http/https handling a bit
The only practical upshot of this really is that we're going
to default to http for localhost if it's explicitly given as
the server; before, explicitly giving 'localhost' as the server
but not specifying a scheme would make the scheme https. This
significantly means you can just have a [localhost] entry on
a server and it'll work properly with both this client and the
perl client, you don't need to have both a [localhost] entry
(for the perl client) and a [http://localhost] entry (for
this client).