iterate_jobs: generator for finished openQA jobs
iterate_jobs is a generator that provides most of the features
of both wait_jobs and wait_build_jobs, and deprecates both of
them. It yields lists of job dicts as they are completed,
rather than waiting for all jobs to complete then returning a
single large list. It also queries for jobs in a more efficient
manner than wait_jobs (made possible by a change to upstream
openQA since wait_jobs was written).

The only missing feature is wait_jobs' ability to operate on
an iterable of job dicts rather than job IDs. Including this
would complicate the new function's code, and we no longer need
that capability for anything (it was introduced for an earlier
design of wait_build_jobs), so it's being removed.

wait_jobs and wait_build_jobs will be removed in a week or so.