allow replacing cloned jobs with their clones
This adds a new method, find_clones(), which accepts an iterable
of job dicts and replaces the dict for any cloned job with the
dict for its clone (recursively - if the clone was itself cloned,
it'll follow the chain till it finds a non-cloned job). It makes
iterate_jobs call find_clones() if filter_dupes is True. The
upshot is you can pass iterate_jobs() a list of job IDs and if
any of those jobs was cloned, it'll wait on the clone rather
than the original (and this is the default behaviour now, since
filter_dupes defaults to True). The code path is also hit when
operating on 'build' rather than 'jobs', but the actual results
are likely not to be different, because filter_dupes would
already happen to achieve the same thing when querying by build.
Where a job has been cloned, the clone would be considered the
'latest' job by get_latest_jobs(). This code is really only
useful for the 'jobs' path, because it actually goes out and
finds the clones (whereas get_latest_jobs() only operates on the
jobs already in the iterable it's passed).