be a bit more proactive with retries in do_request
we had a couple of days in a row where Fedora's check-compose
script bailed in do_request() because (I think) openQA spaced
out for a minute or two, so I thought it'd be a good idea to
get a bit more defensive on the retries. This spaces out retry
attempts a bit more - the first retry is for as many seconds
as 'wait' says, then for each retry, we double the wait period,
up to a max of 60 seconds (and the default is now 10 not 5) -
and retries (rather than bailing immediately) when we get a
ConnectionError from requests. Hopefully this will do the
trick. It also debug logs the exact exception on each failure.