add a simple get_jobs() method with no waiting or iteration
the elaborate waiting/iteration stuff in iterate_jobs() is
built around some Fedora requirements which no longer hold; it
was basically so we could fire up when we knew tests for a
build had started then hang around and wait for them to finish,
doing something with each job result as soon as it appeared. But
now, we have fedmsg! With fedmsg we can do this much better: we
just wait for 'job_done' fedmsgs and fire up every time we see
one. No fuss, no mess. iterate_jobs() is now just an exotic
design whose iteration we have to work around. So let's add a
simpler get_jobs which simply returns the jobs you ask it for
right away, with the intent of removing iterate_jobs so long as
no-one else is using it. All Fedora's users will be moved over
to this method.