add a new module to export some useful openQA constants
I found myself duplicating a lot of openQA job state and result
constants on the fly in openQA-python-client consumers, which
is bad; lots of stuff that can go stale quietly. So I think it's
a good idea to do this. We will have to keep these definitions
in sync with openQA manually, but that's an awful lot easier to
do just once here than do across umpteen consumers of this
package.