**API**: update constants to match upstream 4d89041
Upstream changed the job constants a bit, so we must change to
follow them. The most notable change is that the
JOB_INCOMPLETE_RESULTS tuple is gone. JOB_NOT_COMPLETE_RESULTS
is *NOT* the same, so do not use it as a replacement. However,
JOB_COMPLETE_RESULTS is the exact inverse, so you can trivially
change a check for something being in JOB_NOT_COMPLETE_RESULTS
to a check for it not being in JOB_COMPLETE_RESULTS, for e.g.

Aside from that, new constants were added:

* JOB_NOT_COMPLETE_RESULTS
* JOB_ABORTED_RESULTS
* JOB_NOT_OK_RESULTS

and these new 'meta' states and results:

* JOB_STATE_PRE_EXECUTION
* JOB_STATE_EXECUTION
* JOB_STATE_FINAL

* JOB_RESULT_COMPLETE
* JOB_RESULT_NOT_COMPLETE
* JOB_RESULT_ABORTED

Signed-off-by: Adam Williamson <awilliam@redhat.com>