find_clones: don't edit list while iterating it
This is a classic mistake and means this method has been broken
for a while, probably not returning some jobs it should have
returned. Fortunately it probably hasn't broken anything too
important in practice - fedora_openqa only uses filter_dupes for
wiki reporting, and we usually report wiki results one at a time
as soon as the jobs complete, so this bug won't have mattered. It
*may* have affected fedora_nightlies and check-compose, though.

Signed-off-by: Adam Williamson <awilliam@redhat.com>