Move source under src/ , fix tox config to run tests on package
See https://hynek.me/articles/testing-packaging/ , in general.
We want to build an sdist and run the tests on that, to make
sure our sdist process is good, we don't want to just run the
tests on whatever's in the working dir. You can actually do this
with tox without moving the source under src/ - just removing
skipsdist from tox.ini does that, tox actually wants to do this
by default and I've just been cargo-culting that skipsdist around
for some reason - but if you do that, pytest-cov starts breaking,
I think on this:

https://github.com/pytest-dev/pytest-cov/issues/134#issuecomment-281883741

moving the source under src/ seems to fix that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>