black-ify code and add black to CI config
Let's use the shiny new opinionated code formatter everyone's
talking about:

https://github.com/psf/black

in addition to pylint, for now, because I think pylint may still
catch stuff black doesn't. This reformats the code and tests to
pass black - mostly just replacing ' with " - and adds a black
check to the CI checks. We except const.py because black doesn't
like the padding spaces we use there, and we set the line length
to 100 because that's the line length God intended, don't @ me.

Signed-off-by: Adam Williamson <awilliam@redhat.com>