Handle settings in openqa_request sanely
We have to work around a limitation in the API:
when modifying job groups, products, etc. that take a settings parameter, then
this settings parameter gets returned to us as a list like this:
[{"key": "varname", "value": "var_value"}]

But when we sent the reply back, we must send these settings in the "top level"
payload object like this:
"settings[varname]": "var_value"

The openqa_request method now does this for you, eliminating surprises.