Add class global retries & wait values
The user might not like the default values that we choose for wait and retries,
as this can take a long time until errors actually result in an exception being
raised.
This commit adds class global defaults & values for retries and wait, so that
the user can optionally provide their own defaults that will be used by the
client instance.
However, if the user does not provide them, then nothing changes from their
point of view.

Co-authored-by: Adam Williamson <awilliam@redhat.com>