Add project metadata to pyproject.toml
I'm keeping setup.py for now because the support for this stuff
is only in very new versions of setuptools (61+), so building
packages on older Fedora does not work if we drop setup.py. But
you *can* get rid of setup.py and still build sdists, wheels and
so on if you have setuptools 61 or newer.

Signed-off-by: Adam Williamson <awilliam@redhat.com>