Only retry requests on certain status codes
It doesn't make sense to retry e.g. 404s, or other codes that
would tend to indicate a user error or permanent problem rather
than some transient issue. This is my best cut at a list of
'appropriate' codes to retry on, based on the obvious canonical
source, http.cat.

Signed-off-by: Adam Williamson <awilliam@redhat.com>