openqa_request: correct settings dict conversion
Since the dropping of hashedparams in
https://github.com/os-autoinst/openQA/pull/5994 , this format
has been incorrect. We're now meant to send the settings as a
simple key: value dict (what a concept!), not in the wacky
hashedparams format that allowed sending them as query args *or*
the wacky database-inspired "list of dicts" format. So we need
to convert the wacky list-of-dicts format to a simple dict, not
the hashedparams format.

Signed-off-by: Adam Williamson <awilliam@redhat.com>