CI: Remove obsolete and broken py36
In GitHub as well as tox python3.6 is not properly supported anymore. We
tried to get it to work with custom approaches but to no avail. We have
to disable py36 and handle any compatibility issues downstream as
applicable. Still keeping the support in tox as it can work there with
sufficiently old versions of dependencies we probably have in py3.6
build+test environments.